Stops using Serial.parseInt to resolve a timing issue.
Serial.parseInt waits up to 1s for user input before returning.
aintnobodygottimeforthat.

Also updates trigger polling code slightly to avoid throwing away trigger
events when we run late. Previously if trigcnt >= 10, we make it 10, which
means that if trigcnt is 12, we need another 8 triggers before triggering
again. The new method, which subtracts 10 from tigcnt, means we run on the
20th trigger as expected.

Note of course we expect to never be so late as to miss 2x_DIVIDER number of
triggers, in which case this code triggers a consecutive scans.